feat(httpd): Persist sessions in db with expiration
Add support for `--session-expiry` runtime option, which changes the default authorized session duration to a custom duration. Configuring session expiry to a zero or negative value, makes sessions last indefinitely.

Add support for persisting sessions in (sqlite) db, under `/httpd/sessions.db`, replacing the in-memory map.
Combined with the support for configuring custom session expiration, this allows httpd sessions to be used by other systems, such as CI, with session id used as long-lived "api tokens".
When the session expiration is customized, then the generated session id is doubled in length.
The persisted session id in `sessions.db` is the output hash of the signature on the original session id.
This way, the original session id cannot be recovered from the database.
The sessions db supports a comment column, which allows adding a description for the usage of the generated session.
This comment can be provided when authorizing the session, along with the private key and the signature (by default the comment is empty).
Old/expired sessions are cleaned up when httpd is (re-)started.