moves signing step to inside the container image build
the existing setup is more suited to a development environment and I want to move our build process to a CI environment, so that this is fully automated and all these actions happen in CI jobs.

In a CI environment, it is restrictive to require VMs with docker/podman pre-installed, or to require docker-in-docker setups for CI engines that use containers as workers.

Instead, I want the entire build + signing process to happen inside a container image build - and we can extract the artifacts from the image when done.

This also makes it even clearer what is required wherever this build process will run. (Running `build/build` on a host still doesn't explain what needs to be installed on that host for the script to work).

Signed-off-by: Yorgos Saslis <yorgos.work@proton.me>