Moves signing of binaries inside container image build
this provides an entirely reproducible environment for the whole build+signing process, so that it can be automated and moved to run as part of a pipeline, with a fully automated build-test-release process.

The changes require the keys to be bind-mounted to the container and the passphrase to be passed as an env var to the container image build, so that the key can be unlocked and used for signing the binaries.

Signed-off-by: Yorgos Saslis <yorgos.work@proton.me>